Move walk functions to own file
